fix(slides): clean live deck diagnostics
