+++ src/lib/domains/slides/route-shell.tsx
@@ -16,13 +16,7 @@ import {
 } from "@/lib/domains/slides/live-session";
 import { useSlideSession } from "@/lib/domains/slides/use-slide-session";
 
-export type SlideSearch = {
-	live?: string;
-	master?: boolean;
-	presenter?: boolean;
-	slide?: number;
-	step?: number;
-};
+import { type SlideSearch } from "./route-search";
 
 type SlideModule = {
 	default: Array<Slide>;
@@ -35,16 +29,6 @@ type SlideDeckRouteShellProps = {
 	livePolls: Array<SlideSessionPollDefinition>;
 	onPositionChange: (slide: number, step: number) => void;
 };
-
-export function validateSlideSearch(search: Record<string, unknown>): SlideSearch {
-	return {
-		live: parseSessionId(search.live),
-		master: parseBooleanParam(search.master),
-		presenter: parseBooleanParam(search.presenter),
-		slide: parseNonNegativeInt(search.slide),
-		step: parseNonNegativeInt(search.step),
-	};
-}
 
 export function SlideDeckRouteShell({
 	search,
@@ -161,21 +145,4 @@ function SlideDeckLoader({
 			hide_slide_index
 		/>
 	);
-}
-
-function parseNonNegativeInt(raw: unknown): number | undefined {
-	if (raw === undefined || raw === null || raw === "") return undefined;
-	const n = Number(raw);
-	return Number.isFinite(n) && Number.isInteger(n) && n >= 0 ? n : undefined;
-}
-
-function parseSessionId(raw: unknown): string | undefined {
-	if (typeof raw !== "string") return undefined;
-	return /^[a-zA-Z0-9_-]{1,80}$/.test(raw) ? raw : undefined;
-}
-
-function parseBooleanParam(raw: unknown): boolean | undefined {
-	if (raw === "1" || raw === "true" || raw === 1 || raw === true) return true;
-	if (raw === "0" || raw === "false" || raw === 0 || raw === false) return false;
-	return undefined;
 }
